Add files via upload
The meaning of this project code is to create a web page that shows a simulation of the planet Mars and some rockets and rocks. The web page uses HTML to structure the content, CSS to style the appearance, and JavaScript to add animation and interaction. The web page is an example of a creative and innovative project that can be done with HTML and other web technologies.